Adding a button and a delay
So now we've got a button that will start the game once pressed. The game basically makes a colored block appear in an interval between two and three seconds. The main thing left to implement now is the reaction timer itself, which should show how much time a user spent on click the block after it's appeared